OSN-1260. Rename Components folder to components
