--- ngui/ui/src/components/BIExport/Components/index.ts
@@ -1,4 +1,0 @@
-import FilesSummaryList from "./FilesSummaryList";
-import TargetStorageSummaryList from "./TargetStorageSummaryList";
-
-export { FilesSummaryList, TargetStorageSummaryList };
