+++ ngui/ui/src/components/BIExport/components/index.ts
@@ -1,0 +1,5 @@
+import DetailsSummaryList from "./DetailsSummaryList";
+import FilesSummaryList from "./FilesSummaryList";
+import TargetStorageSummaryList from "./TargetStorageSummaryList";
+
+export { FilesSummaryList, TargetStorageSummaryList, DetailsSummaryList };
