+++ ngui/ui/src/components/MlRunsetOverview/components/index.ts
@@ -1,5 +1,6 @@
+import Correlations from "./Correlations";
 import InputParameters from "./InputParameters";
 import SummaryCards from "./SummaryCards";
 import Tabs from "./Tabs";
 
-export { Tabs, InputParameters, SummaryCards };
+export { Tabs, InputParameters, SummaryCards, Correlations };
