+++ ngui/ui/src/components/MlRunsetTemplate/components/Details.tsx
