+++ ngui/ui/src/components/RecommendationsCard/RecommendationsCard.tsx
@@ -11,7 +11,7 @@ import { CATEGORY } from "containers/RecommendationsOverviewContainer/recommenda
 import { ALL_SERVICES } from "hooks/useRecommendationServices";
 import { RECOMMENDATIONS, RECOMMENDATION_CATEGORY_QUERY_PARAMETER, RECOMMENDATION_SERVICE_QUERY_PARAMETER } from "urls";
 import { SPACING_2 } from "utils/layouts";
-import { InfoCard, PossibleSavingsCard } from "./Components";
+import { InfoCard, PossibleSavingsCard } from "./components";
 
 const RecommendationsCard = ({
   isLoading,
