+++ ngui/ui/src/components/RecommendationsCard/components/InfoCard/InfoCard.tsx
