+++ ngui/ui/src/components/RecommendationsCard/components/InfoCard/index.ts
