+++ ngui/ui/src/components/RecommendationsCard/components/PossibleSavingsCard/index.ts
